fix(file-controller): resourceId are not always UUIDs

+++ libs/feature/resource/server/src/lib/files/file.controller.ts
@@ -19,9 +19,9 @@ import { ConfigService } from '@nestjs/config'
 import { FileInterceptor } from '@nestjs/platform-express'
 import { ApiTags } from '@nestjs/swagger'
 import { BadRequestResponse, SuccessResponse, UnauthorizedResponse } from '@platon/core/common'
-import { Configuration, EventService, IRequest, Public, UUIDParam } from '@platon/core/server'
+import { Configuration, EventService, IRequest, Public } from '@platon/core/server'
 import { PLSourceFile } from '@platon/feature/compiler'
-import { ExerciseTransformInput, FileTypes, LATEST, ResourceFile } from '@platon/feature/resource/common'
+import { ExerciseTransformInput, FileTypes, LATEST, ResourceFile, ResourceTypes } from '@platon/feature/resource/common'
 import { Response } from 'express'
 import * as fs from 'fs'
 import mime from 'mime-types'
@@ -97,7 +97,7 @@ export class ResourceFileController {
   ) {}
 
   @Get('/log/:resourceId')
-  async log(@Req() request: IRequest, @UUIDParam('resourceId') resourceId: string) {
+  async log(@Req() request: IRequest, @Param('resourceId') resourceId: string) {
     const { repo, permissions } = await this.fileService.repo(resourceId, request)
     if (!permissions.read) {
       throw new UnauthorizedResponse('You are not allowed')
@@ -106,7 +106,7 @@ export class ResourceFileController {
   }
 
   @Post('/release/:resourceId')
-  async release(@Req() request: IRequest, @UUIDParam('resourceId') resourceId: string, @Body() input: FileReleaseDTO) {
+  async release(@Req() request: IRequest, @Param('resourceId') resourceId: string, @Body() input: FileReleaseDTO) {
     const { repo, resource, permissions } = await this.fileService.repo(resourceId, request)
     if (!permissions.write) {
       throw new UnauthorizedResponse('You are not allowed to release this resource')
@@ -125,7 +125,7 @@ export class ResourceFileController {
   @Post('/compile/:resourceId/json')
   async compileExercise(
     @Req() request: IRequest,
-    @UUIDParam('resourceId') resourceId: string,
+    @Param('resourceId') resourceId: string,
     @Query('version') version = LATEST
   ): Promise<PLSourceFile> {
     const { source } = await this.fileService.compile({ resourceId, version, req: request, withAst: true })
@@ -135,7 +135,7 @@ export class ResourceFileController {
   @Post('/compile/:resourceId/text')
   async transformExercise(
     @Req() request: IRequest,
-    @UUIDParam('resourceId') resourceId: string,
+    @Param('resourceId') resourceId: string,
     @Query('version') version = LATEST,
     @Body() input?: ExerciseTransformInput
   ): Promise<string> {
@@ -266,7 +266,7 @@ export class ResourceFileController {
   )
   async put(
     @Req() request: IRequest,
-    @UUIDParam('resourceId') resourceId: string,
+    @Param('resourceId') resourceId: string,
     @Param('path') path: string,
     @Body() input: FileUpdateDTO,
     @UploadedFile() bundle: Express.Multer.File
@@ -316,7 +316,7 @@ export class ResourceFileController {
   )
   async post(
     @Req() request: IRequest,
-    @UUIDParam('resourceId') resourceId: string,
+    @Param('resourceId') resourceId: string,
     @Body() input: FileCreateDTO[],
     @UploadedFile() file: Express.Multer.File,
     @Param('path') path?: string
@@ -366,7 +366,7 @@ export class ResourceFileController {
   @Patch('/:resourceId/:path(*)')
   async patch(
     @Req() request: IRequest,
-    @UUIDParam('resourceId') resourceId: string,
+    @Param('resourceId') resourceId: string,
     @Param('path') path: string,
     @Body() input: FileMoveDTO
   ) {
@@ -410,10 +410,13 @@ export class ResourceFileController {
   }
 
   @Delete('/:resourceId/:path(*)')
-  async delete(@Req() request: IRequest, @UUIDParam('resourceId') resourceId: string, @Param('path') path: string) {
+  async delete(@Req() request: IRequest, @Param('resourceId') resourceId: string, @Param('path') path: string) {
     const { repo, resource, permissions } = await this.fileService.repo(resourceId, request)
     if (!permissions.write) {
       throw new UnauthorizedResponse('You are not allowed to write this resource')
+    }
+    if (resource.type === ResourceTypes.CIRCLE && request.user.role !== 'admin') {
+      throw new UnauthorizedResponse('You are not allowed to delete this resource')
     }
 
     await repo.remove(path)
